// app/invoice-app/src/app/api-client/model/customer.ts
/**
 * Invoice API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Customer { 
    id?: number;
    name: string;
    email: string;
    contactPerson?: string;
    location?: string;
    postalCode?: string;
    street?: string;
    mainNumber?: string;
}